Implemented AEC frame data format and UDP transport
Support sending valid frames at a rate of 25 FPS
to the facade. The idea is to poll for display
buffer updates and always send the latest buffer
content as the current frame. This will support an
API where only deltas between last and current frame
need to be transmitted.

Added basic unit specs for the display buffer and
a simple integration spec that uses a faked aec
facade server.

To run the specs, you need to have rake and rspec
installed. Once you have that, run 'rake spec' in
the toplevel project directory. You should see
5 passing, 0 failing, and 1 pending spec.