Introduced display controller to handle walls and screens
The display controller will be responsible for managing
all different walls and screens on these walls. Walls
are basically the different parts of the AEC facades,
whereas screens are logical units that are managed within
(or spread over more than one) walls.

Layers above the display manager can use it to both to
(un)register screens on the display, and to update the
display either as a whole, or on a per pixel basis.